+++ Package.swift
@@ -5,13 +5,13 @@ import PackageDescription
 
 let package = Package(
     name: "Reach5Google",
-    platforms: [.iOS(.v13), .macOS(.v10_15)],
+    platforms: [.iOS(.v13)],
     products: [
         .library(name: "Reach5Google", targets: ["Reach5Google"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/ReachFive/reachfive-ios.git", "9.0.0"..<"100.0.0"),
-        .package(url: "https://github.com/google/GoogleSignIn-iOS.git", .upToNextMajor(from: "7.1.0")),
+        .package(url: "https://github.com/ReachFive/reachfive-ios.git", .upToNextMajor(from: "11.0.0")),
+        .package(url: "https://github.com/google/GoogleSignIn-iOS.git", .upToNextMajor(from: "9.2.0")),
     ],
     targets: [
         .target(